Core: added support for the new 4.1 combat log argument "hideCaster"
this is a temporary work-around which just drops the new argument for a quick and easy fix. It doesn't look like that we are going to need this argument ever, so this is okay for now.
TODO: apply this change to the actual function above when 4.1 goes live, hideCaster can then be added to the args table

git-svn-id: svn://dev.deadlybossmods.com/dbm4/trunk@5260 a6602f14-c1bc-49d1-8d65-da3458cc0765