Implement fast event processing for high volume events (like SPELL_DAMAGE), these new events will no longer be using the DBM args table but the raw event arguments All events that are changed can be found in DBM-Core.lua (see diff)
The reason for this change is that old boss mods didn't really need these events, so performance issues with these high volume events were no issue. However, a lot of newer mods make use of these events for some hacks. The usual approach for the few mods which actually needed these events was to build a new CLEU event handler to bypass DBM event processing. This was reasonable as long as there were like 2 or 3 mods that did this. But nowadays there are like 3 oder 4 mods in each package that require these high volume events...

Note that this change is incompatible with old mods. All mods that were using these events will require updating. See DragonSoul/SpineDeathwing for an example.



git-svn-id: svn://dev.deadlybossmods.com/dbm4/trunk@7259 a6602f14-c1bc-49d1-8d65-da3458cc0765