tandanu via svn r11848: svn server had a disk failure and crashed. that somehow corrupted the metadata for the last few commits (thanks, svn!)
the simplest way to fix this seemed to be to apply the diffs for these commits by hand and remove the bad revision from the server

here are the old commit messages:
Log:
Improve note, since I actually got not one but "two" bug reports about getting "taunt" warnings when not tank spec, when disabling this filter. Um that's the point?. An on by default filter and on by default taunt warnings allows tanks that very frequently switch between dps/tank to not have to worry about if they remembered to enable/disable the taunt alerts. They "just work". don't complain if you disable the filter and keep getting told to taunt the boss. :P

Log:
# clean.

Log:
Fix missing icon